requirements.txt
# razomy.function.server
uvicorn~=0.30.0
fastapi~=0.111.0
jinja2~=3.1.4
# razomy.python.image
pillow~=10.3.0
